+++ src/main/kotlin/com/example/custard/domain/common/date/DateStore.kt
@@ -1,7 +1,6 @@
-package com.example.custard.domain.post.service.date
+package com.example.custard.domain.common.date
 
-import com.example.custard.domain.post.dto.info.DateInfo
-import com.example.custard.domain.post.model.date.Date
+import com.example.custard.domain.common.date.dto.DateInfo
 
 interface DateStore {
     fun saveDateIfNotExists(date: Date): Date
