+++ src/main/kotlin/com/example/custard/domain/common/date/DateStoreImpl.kt
@@ -1,8 +1,6 @@
-package com.example.custard.domain.post.repository.date;
+package com.example.custard.domain.common.date;
 
-import com.example.custard.domain.post.dto.info.DateInfo
-import com.example.custard.domain.post.model.date.Date
-import com.example.custard.domain.post.service.date.DateStore
+import com.example.custard.domain.common.date.dto.DateInfo
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
