+++ src/main/kotlin/com/example/custard/domain/post/model/Post.kt
@@ -1,6 +1,5 @@
 package com.example.custard.domain.post.model
 
-import com.example.custard.domain.post.model.date.PostDate
 import com.example.custard.domain.user.model.User
 import jakarta.persistence.*
