+++ src/main/kotlin/com/example/custard/domain/post/repository/date/PostDateRepository.kt
@@ -1,8 +1,8 @@
 package com.example.custard.domain.post.repository.date
 
 import com.example.custard.domain.post.model.Post
-import com.example.custard.domain.post.model.date.Date
-import com.example.custard.domain.post.model.date.PostDate
+import com.example.custard.domain.common.date.Date
+import com.example.custard.domain.post.model.PostDate
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PostDateRepository : JpaRepository<PostDate, Long> {
